AT_AFO_project/requirements.txt
numpy
pandas
matplotlib
pyserial
scipy
scikit-learn
